DDCNL-9021 | moved /confirm into its own controller & tests
